20150512
更新了互联网的喜欢的公司的存储问题, 除了地图没有弄之外,项目基本完成

+++ ZYMoreWork/视图/找工作/报名/BaomingViewController.m
@@ -84,7 +84,7 @@
 }
 -(void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex
 {
-    NSLog(@"buttonIndex%u",buttonIndex);
+    NSLog(@"buttonIndex%ld",(long)buttonIndex);
     if (buttonIndex==0) {
         NSLog(@"知道了");
     }
